requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.2
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==23.1.0
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
bleach==6.0.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.3
cron-descriptor==1.4.0
cryptography==40.0.2
customtkinter==5.1.3
darkdetect==0.8.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2.1
django-celery-beat==2.5.0
django-constance==2.9.1
django-cors-headers==4.0.0
django-migration-linter==4.1.0
django-model-utils==4.3.1
django-picklefield==3.1
django-query-counter==0.2.2
django-timezone-field==5.0
djangorestframework==3.14.0
dnspython==2.4.2
envparse==0.2.0
exceptiongroup==1.1.3
executing==1.2.0
Faker==12.0.1
fastapi==0.81.0
fastapi-utils==0.2.1
fastjsonschema==2.16.3
filelock==3.12.2
fqdn==1.5.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.1.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.22.1
kombu==5.3.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
mixer==7.2.2
motor==3.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
opencv-python==4.5.4.60
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.3
pdf2image==1.16.0
pdfminer.six==20221105
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
Pygments==2.15.1
pymongo==4.5.0
PyMuPDF==1.22.1
PyNaCl==1.5.0
pyrsistent==0.19.3
pytest==6.2.5
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.10
pyxel==1.9.16
PyYAML==6.0
pyzbar==0.1.9
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
redis==4.5.5
requests==2.30.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.19.1
streamz==0.6.4
tabulate==0.9.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.2.0
urllib3==2.0.2
uvicorn==0.18.3
vine==5.0.0
virtualenv==20.23.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
widgetsnbextension==4.0.7
zict==3.0.0
zipp==3.15.0
